Crate serialize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Support code for encoding and decoding types.

Modules

hex 	 [
Experimental
]
Hex binary-to-text encoding

json 	 [
Experimental
]
JSON parsing and serialization

leb128 	 [
Experimental
]

opaque 	 [
Experimental
]

Traits

Decodable 	 [
Experimental
]

Decoder 	 [
Experimental
]

Encodable 	 [
Experimental
]

Encoder 	 [
Experimental
]

SpecializationError 	 [
Experimental
]
Implement this trait on your {Encodable,Decodable}::Error types to override the default panic behavior for missing specializations.

SpecializedDecoder 	 [
Experimental
]
Implement this trait on decoders, with T being the type you want to decode (employing UseSpecializedDecodable), using a strategy specific to the decoder.

SpecializedEncoder 	 [
Experimental
]
Implement this trait on encoders, with T being the type you want to encode (employing UseSpecializedEncodable), using a strategy specific to the encoder.

UseSpecializedDecodable 	 [
Experimental
]
Implement this trait on your type to get an Decodable implementation which goes through SpecializedDecoder.

UseSpecializedEncodable 	 [
Experimental
]
Implement this trait on your type to get an Encodable implementation which goes through SpecializedEncoder.