Trait serialize::hex::FromHex

pub trait FromHex {
    fn from_hex(&self) -> Result<Vec<u8>, FromHexError>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A trait for converting hexadecimal encoded values

Required Methods

fn from_hex(&self) -> Result<Vec<u8>, FromHexError>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Converts the value of self, interpreted as hexadecimal encoded data, into an owned vector of bytes, returning the vector.