Trait serialize::hex::ToHex

pub trait ToHex {
    fn to_hex(&self) -> String;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A trait for converting a value to hexadecimal encoding

Required Methods

fn to_hex(&self) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Converts the value of self to a hex value, returning the owned string.