Enum serialize::json::Json

pub enum Json {
    I64(i64),
    U64(u64),
    F64(f64),
    String(String),
    Boolean(bool),
    Array(Array),
    Object(Object),
    Null,
}

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Represents a json value

Variants

I64
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

U64
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

F64
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

String
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Boolean
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Array
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Object
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Null
Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Methods

impl Json

fn pretty(&self) -> PrettyJson

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Borrow this json object as a pretty object to generate a pretty representation for it via Display.

fn find<'a>(&'a self, key: &str) -> Option<&'a Json>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is an Object, returns the value associated with the provided key. Otherwise, returns None.

fn find_path<'a>(&'a self, keys: &[&str]) -> Option<&'a Json>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Attempts to get a nested Json Object for each key in keys. If any key is found not to exist, find_path will return None. Otherwise, it will return the Json value associated with the final key.

fn search<'a>(&'a self, key: &str) -> Option<&'a Json>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is an Object, performs a depth-first search until a value associated with the provided key is found. If no value is found or the Json value is not an Object, returns None.

fn is_object<'a>(&'a self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is an Object. Returns false otherwise.

fn as_object<'a>(&'a self) -> Option<&'a Object>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is an Object, returns the associated BTreeMap. Returns None otherwise.

fn is_array<'a>(&'a self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is an Array. Returns false otherwise.

fn as_array<'a>(&'a self) -> Option<&'a Array>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is an Array, returns the associated vector. Returns None otherwise.

fn is_string<'a>(&'a self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a String. Returns false otherwise.

fn as_string<'a>(&'a self) -> Option<&'a str>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a String, returns the associated str. Returns None otherwise.

fn is_number(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a Number. Returns false otherwise.

fn is_i64(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a i64. Returns false otherwise.

fn is_u64(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a u64. Returns false otherwise.

fn is_f64(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a f64. Returns false otherwise.

fn as_i64(&self) -> Option<i64>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a number, return or cast it to a i64. Returns None otherwise.

fn as_u64(&self) -> Option<u64>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a number, return or cast it to a u64. Returns None otherwise.

fn as_f64(&self) -> Option<f64>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a number, return or cast it to a f64. Returns None otherwise.

fn is_boolean(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a Boolean. Returns false otherwise.

fn as_boolean(&self) -> Option<bool>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a Boolean, returns the associated bool. Returns None otherwise.

fn is_null(&self) -> bool

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Returns true if the Json value is a Null. Returns false otherwise.

fn as_null(&self) -> Option<()>

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

If the Json value is a Null, returns (). Returns None otherwise.