Struct serialize::json::Decoder

pub struct Decoder { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A structure to decode JSON to values in rust.

Methods

impl Decoder

pub fn new(json: Json) -> Decoder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Creates a new decoder instance for decoding the specified JSON value.