Struct serialize::json::PrettyEncoder

pub struct PrettyEncoder<'a> { /* fields omitted */ }

Another encoder for JSON, but prints out human-readable JSON instead of compact data

Methods

impl<'a> PrettyEncoder<'a>

pub fn new(writer: &'a mut dyn Write) -> PrettyEncoder<'a>

Creates a new encoder whose output will be written to the specified writer

pub fn set_indent(&mut self, indent: usize)

Sets the number of spaces to indent for each level. This is safe to set during encoding.