Struct serialize::json::Stack

pub struct Stack {
    stack: Vec<InternalStackElement>,
    str_buffer: Vec<u8>,
}

A Stack represents the current position of the parser in the logical structure of the JSON stream. For example foo.bar[3].x

Fields
stack: Vec<InternalStackElement> str_buffer: Vec<u8>
Methods

impl Stack

pub fn new() -> Stack

pub fn len(&self) -> usize

Returns The number of elements in the Stack.

pub fn is_empty(&self) -> bool

Returns true if the stack is empty.

pub fn get(&self, idx: usize) -> StackElement

Provides access to the StackElement at a given index. lower indices are at the bottom of the stack while higher indices are at the top.

pub fn is_equal_to(&self, rhs: &[StackElement]) -> bool

Compares this stack with an array of StackElements.

pub fn starts_with(&self, rhs: &[StackElement]) -> bool

Returns true if the bottom-most elements of this stack are the same as the ones passed as parameter.

pub fn ends_with(&self, rhs: &[StackElement]) -> bool

Returns true if the top-most elements of this stack are the same as the ones passed as parameter.

pub fn top(&self) -> Option<StackElement>

Returns the top-most element (if any).

fn push_key(&mut self, key: String)

fn push_index(&mut self, index: u32)

fn pop(&mut self)

fn last_is_index(&self) -> bool

fn bump_index(&mut self)