Trait serialize::SpecializationError

pub trait SpecializationError {
    fn not_found<S, T: ?Sized>(trait_name: &'static str,
                               method_name: &'static str)
                               -> Self;
}

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Implement this trait on your {Encodable,Decodable}::Error types to override the default panic behavior for missing specializations.

Required Methods

fn not_found<S, T: ?Sized>(trait_name: &'static str,
method_name: &'static str)
-> Self

Unstable (rustc_private)
: deprecated in favor of rustc-serialize on crates.io

Create an error for a missing method specialization. Defaults to panicking with type, trait & method names. S is the encoder/decoder state type, T is the type being encoded/decoded, and the arguments are the names of the trait and method that should've been overriden.