Trait serialize::SpecializedEncoder

pub trait SpecializedEncoder<T: ?Sized + UseSpecializedEncodable>: Encoder {
    fn specialized_encode(&mut self, value: &T) -> Result<(), Self::Error>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Implement this trait on encoders, with T being the type you want to encode (employing UseSpecializedEncodable), using a strategy specific to the encoder.

Required Methods

fn specialized_encode(&mut self, value: &T) -> Result<(), Self::Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Encode the value in a manner specific to this encoder state.