Enum serialize::json::Json

pub enum Json {
    I64(i64),
    U64(u64),
    F64(f64),
    String(String),
    Boolean(bool),
    Array(Array),
    Object(Object),
    Null,
}

Represents a json value

Variants
I64(i64)U64(u64)F64(f64)String(String)Boolean(bool)Array(Array)Object(Object)Null
Methods

impl Json

pub fn pretty(&self) -> PrettyJson

Borrow this json object as a pretty object to generate a pretty representation for it via Display.

pub fn find<'a>(&'a self, key: &str) -> Option<&'a Json>

If the Json value is an Object, returns the value associated with the provided key. Otherwise, returns None.

pub fn find_path<'a>(&'a self, keys: &[&str]) -> Option<&'a Json>

Attempts to get a nested Json Object for each key in keys. If any key is found not to exist, find_path will return None. Otherwise, it will return the Json value associated with the final key.

pub fn search<'a>(&'a self, key: &str) -> Option<&'a Json>

If the Json value is an Object, performs a depth-first search until a value associated with the provided key is found. If no value is found or the Json value is not an Object, returns None.

pub fn is_object(&self) -> bool

Returns true if the Json value is an Object.

pub fn as_object(&self) -> Option<&Object>

If the Json value is an Object, returns the associated BTreeMap; returns None otherwise.

pub fn is_array(&self) -> bool

Returns true if the Json value is an Array.

pub fn as_array(&self) -> Option<&Array>

If the Json value is an Array, returns the associated vector; returns None otherwise.

pub fn is_string(&self) -> bool

Returns true if the Json value is a String.

pub fn as_string(&self) -> Option<&str>

If the Json value is a String, returns the associated str; returns None otherwise.

pub fn is_number(&self) -> bool

Returns true if the Json value is a Number.

pub fn is_i64(&self) -> bool

Returns true if the Json value is a i64.

pub fn is_u64(&self) -> bool

Returns true if the Json value is a u64.

pub fn is_f64(&self) -> bool

Returns true if the Json value is a f64.

pub fn as_i64(&self) -> Option<i64>

If the Json value is a number, returns or cast it to a i64; returns None otherwise.

pub fn as_u64(&self) -> Option<u64>

If the Json value is a number, returns or cast it to a u64; returns None otherwise.

pub fn as_f64(&self) -> Option<f64>

If the Json value is a number, returns or cast it to a f64; returns None otherwise.

pub fn is_boolean(&self) -> bool

Returns true if the Json value is a Boolean.

pub fn as_boolean(&self) -> Option<bool>

If the Json value is a Boolean, returns the associated bool; returns None otherwise.

pub fn is_null(&self) -> bool

Returns true if the Json value is a Null.

pub fn as_null(&self) -> Option<()>

If the Json value is a Null, returns (); returns None otherwise.