Struct serialize::json::PrettyEncoder

pub struct PrettyEncoder<'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Another encoder for JSON, but prints out human-readable JSON instead of compact data

Methods

impl<'a> PrettyEncoder<'a>

pub fn new(writer: &'a mut Write) -> PrettyEncoder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Creates a new encoder whose output will be written to the specified writer

pub fn set_indent(&mut self, indent: usize)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Set the number of spaces to indent for each level. This is safe to set during encoding.